refactor(settings): split the settings sidebar into workspace and org variants [ENG-1705]
Follow-up to the selector removal. The workspace-agnostic and in-workspace
settings routes already render through different shells, but both still went
through one 494-line sidebar body that decided at runtime — and then, after
the previous commit, at the type level — whether to draw the Workspace
section. That conditional is now gone: the two surfaces get their own
component over a shared core.

New modules/settings/components/sidebar/:

- settings-nav-link, settings-section-header, settings-nav-section and
  types — the shared primitives, moved verbatim. settings-nav-section
  replaces an inline renderSection closure and reads pathname itself
  instead of having it threaded through.
- organization-and-account-sections — the Organization and Account blocks
  both sidebars end with, so neither copy owns those item arrays.
- workspace-settings-sidebar — Workspace section plus that tail. Every
  piece of workspace data is required; there is no "hide it" mode.
- organization-settings-sidebar — that tail only. It has no concept of a
  workspace at all.

SettingsSidebarContent and the hideWorkspaceSection prop are deleted, which
also corrects the dependency direction: modules/settings no longer reaches
into app/(app)/workspaces/[workspaceId]/components for its sidebar.

Pure structural change — the rendered DOM is unchanged. The moved component
bodies and all three nav item arrays were byte-compared against the original,
and the shared tail is a fragment so section nesting stays identical. The
ENG-1705 Playwright spec passes with no assertion changes; only its header
comment was updated, since it described the component this commit removes.
